ref: currDeckChoose component
