+++ client/src/components/LoadScene/LoadScene.ts
@@ -19,7 +19,6 @@ export class LoadScene extends Phaser.Scene {
   }
 
   preload(): void {
-    console.log('sc=', this.nextScene);
     preload(this, this.nextScene);
   }
 }
